# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Retail app Api",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "deendipo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://retail-app.test/api",
            "description": " API Documentation Server"
        }
    ],
    "paths": {
...
        "/user/view_inventories": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get list of Inventories for user",
                "description": "Returns list of Inventories for user",
                "operationId": "getInventoryListForUser",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "bolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Package One"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "example": "20000"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2022-07-10T19:16:29.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2022-07-10T19:16:29.000000Z"
                                                    },
                                                    "inventory_quantity": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "number",
                                                                "example": "18"
                                                            },
                                                            "inventory_id": {
                                                                "type": "number",
                                                                "example": "1"
                                                            },
                                                            "quantity": {
                                                                "type": "number",
                                                                "example": "200"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "example": "2022-07-10T19:16:29.000000Z"
                                                            },
                                                            "updated_at": {
                                                                "type": "string",
                                                                "example": "2022-07-10T19:16:29.000000Z"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}